☕️ 🚀[FIX FEATURE][ UPDATE SOME CODE AND FIX MANY TO MANY RELATIONSHIP BETWEEN APPLICANT AND JOB_OFFER ...  ]

+++ src/main/java/ma/yc/api/myrhapi/repository/JobApplicationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface
-JobApplicationRepository extends JpaRepository<JobApplication,Long> {
+JobApplicationRepository extends JpaRepository<JobApplication,Integer> {
     public List<JobApplication> findByStatus(String status );
 
 }
